+++ Examples/ObjC/Sources/Common/ReadmeViewController.h
@@ -4,6 +4,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface ReadmeViewController : UITableViewController
 
+/// Clears the text of all text field rows.
+- (void)clearTextFields;
+
 @end
 
 NS_ASSUME_NONNULL_END