+++ Sources/BrazeUI/InAppMessageUI/Views/InAppMessageUIModalImageView.swift
@@ -344,6 +344,7 @@ extension BrazeInAppMessageUI {
 
         // Close button
         closeButton.anchors.height.equal(closeButton.anchors.width)
+        closeButton.anchors.width.greaterThanOrEqual(Self.closeButtonMinimumTouchTargetSize)
         closeButton.anchors.edges.pin(
           to: contentView.layoutMarginsGuide,
           alignment: .topTrailing